Spanish Translation (WIP)
I have translated all strings from strings.resx into Spanish. However, I am not a native speaker, so I can not vouch for the validity of the translations. Maybe the translation could be flagged as "beta" in the dropdown? 

Also, Spanish is much more verbose than English or German, so some autosize widgets are too small for the strings and will have to be adjusted.